fix: check for unlimited dimension on variable before adjusting counts
A netcdf file may have an unlimited dimension, but a particular variable may not have the unlimited
dimension.
See #519 for CESM2.3 CLM which has the unlimited cohort dimension.  In the example restart and history files
no state variables (or any variables) have the cohort dimension.
Massive assumption in direct_netcdf_mod.f90 that we want to read the one slice (the latest slice) of the unlimited dimension. This may not be true for all models.