start of scripting template to test at the program level
make a shell script that runs a series of programs (in this case
the obs_sequence_tool).  for each test, an input.nml, an input
obs_seq file and an expected output obs_seq file are needed.

the script should run each in turn and print FAIL and exit if
either the tool fails or if the results don't compare.  note that
the compare is a simple diff so running this on different architectures
may result in false failures.  it may need an example output file
that's architecture dependent.  sigh.

right now this has only a single test which runs the obs_seq tool
without any settings, so it should copy the input to the output
unchanged.  more tests can be added by supplying 3 files in the
seqtool_data directory.

comments welcome on whether this is a viable strategy for doing
testing at the program level.